Use Bellman-Ford not Dijkstra
So that we can report whole paths which would cause the cycle.

This will slow down cycle detection a little; if it becomes a problem,
we could do a Dijkstra run, and only if we detect a cycle (which is the
rare case), do the Bellman-Ford. But we've also been talking about
trying to do incremental cycle detection, so I'm not going to worry too
much about this unless it starts posing a noticeable problem.