Replace the `goals` goal with a `help` plugin which supports both v1 and v2 (#7598)
### Problem

In #6880, there will be two types of Goals:
1. v1 `pants.goal.goal.Goal`s, registered on a singleton
2. v2 `pants.engine.goal.Goal`s, registered as a `ScopeInfo` category via `@console_rules`.

But `./pants goals` supports only the first form (via access to the singleton), and would need to dig deep into `self.context` in order to get access to the scope info (...like `./pants options` does: but that's a story for another day!).

### Solution

Convert `./pants goals` into a plugin to the arg splitter, similar to `./pants help`. While the arg-splitter method is not incredibly scalable, it seems like it should be sufficiently scalable to support most of our "meta-goals". 

### Result

v1 and v2 Goals are supported, we have one fewer v1 Task, and #6880 is unblocked.