Add support for v2-only goals, and replace list with a @console_rule (#6880)
### Problem

While `@console_rule` has existed for a while now, it has not yet been possible to install a `@console_rule` as the exclusive implementer of a goal like `./pants list` (see #6918 and #6651). Additionally, replacing existing goals comes with a need to backwards-compatibly consume options.

### Solution

Run goals that are unambiguously `--v1` or `--v2` without passing those flags: this means that passing the `--v1` vs `--v2` flags is only necessary in cases where both `v1` and `v2` goals are installed.  Additionally, create an `engine.goal.Goal` class with an inner `Optionable` instance in order to claim a CLI scope (and have options), and require one for all `@console_rules`. Finally, add a `ConsoleRuleTestBase` class to easily unit test `@console_rules`.

### Result

The `list` goal is now implemented as a `@console_rule`, and consequently: is faster. `./pants help list` is unchanged due to a deprecated dependency on a `CacheSetup` subsystem, and an optional `Goal`-mixin for wrapping line-oriented output around a `Console`. Fixes #6918 and fixes #6651.