Pin OSX wheel release shards to pyenv. (#7591)
Previously we only did this for the UCS4 OSX shard which allowed the
system interpreters to leak in to interpreter discovery on the other
shards. This was problematic since the system interpreters were linked
against too-old openssl leading to HTTPS errors attempting to talk to
pypi with too-old TLS protocols.

Also switch away from pyenv shims to the explicit version we need for
maximum clarity.

Finally, ensure failures running packages.py propagate. The fact that
they did not led to the issue here going undiscovered in branch CI.

Fixes #7593