pantsd auto invalidates pants.ini and all pythonpath of pants (#7599)
The invalidation globs for pantsd (improved in #5567) can be manually configured to cover everything that is known to cause pantsd to need to restart. But many of the things that should trigger a restart are already known to pants. A partial list:
  1. The content of any configured pants.ini files (due to #7022)
  2. The pythonpath of pants itself (since changes to loose-source plugins mean the code that pantsd is running might have changed)

We should automatically include these values (and likely others!) in the values that we use for --invalidation-globs. In cases where the options values point to files that exist outside of the buildroot, we should consider logging a warning (unless that ends up being too noisy).

Fixes #7595.