Default release script to use Python 3 in preparation for dropping Python 2 release (#7608)
In 1.17.0.dev0, we will no longer release Python 2 wheels or PEX. This is prework to make for a smaller diff when removing that support.

Beyond defaulting to Py3 in `release.sh` and `packages.py`, we now build fs util with the Py36 shards. This is now possible because we now build the Py36 engine during the cron run thanks to Pantsd testing.