PythonTests force default platform resolves
Fixes #7616.

Without this PR, if a test and a binary are both on the same command
line, and the binary specifies an explicit platform not in the default
platforms, we won't resolve for the default platform, and the test will
fail to run because it will be missing dependencies for its platform.

This forces any PythonTests target in the graph to contribute to which
platforms will be resolved.