unify precedence logic for options which may be overridden on targets (#7594)
### Problem

Addresses the first item of #7183. `MirroredTargetOptionMixin` was created in `backend/native/` to make it easier to support pants options with equivalents on a target payload entry, for example. This generalizes that and uses it to simplify some options in tasks in the jvm backend.

### Solution

*commits are independently reviewable*
- move `MirroredTargetOptionMixin` to `build_graph/` and substantially refactor it.
- make `ZincLanguageMixin` a `MirroredTargetOptionMixin` (this allows some of the `DependencyContext` logic to be simplified).
- remove the `--default-workflow` option from `RscCompile`, renaming it to `--workflow` and using the shared logic from `MirroredTargetOptionMixin`.

### Result

The path to create options which can be overridden by targets and to access them from a task is more clear. The RscCompile task now uses the shared option precedence logic of MirroredTargetOptionMixin instead of reinventing it.