Add testing for improved error message from #7628 when no valid interpreter can be resolved (#7630)
While we currently do check that the branch for no valid interpreter being detected works, the check is not very comprehensive. 

In https://github.com/pantsbuild/pants/pull/7628, we made the logged error message even more complex, so it becomes even more important to check that we are logging the error message correctly.