allow configuration of the subsystems to initialize in a test (#7665)
### Problem

Some target types, such as `PythonBinary`, depend on subsystems which they access during initialization. To test instantiation of these targets, their subsystems need to be initialized. We already do this with the subsystems for the `Target` base class -- this just exposes that to test classes.

### Solution

- Introduce the `TestBase.subsystems` classproperty.
- Add testing demonstrating its use.

### Result

Tests which instantiate targets that require subsystems to be initialized can declare that with a `classproperty`.