Short-circuit primitive encoding on CoercingEncoder (#7655)
### Problem

`CoercingEncoder.encode()` is called many times when fingerprinting options. There is a significant overhead to the `isinstance()` calls in python [reference](https://stackoverflow.com/questions/42378726/why-is-checking-isinstancesomething-mapping-so-slow). This means that, when `datatype`s recursively encode themselves after #7304, they do twice the checks that they already did, causing a several-second regression in large targets (from 56s to 59s, without pantsd).

### Solution

Short-circuit primitives to avoid some of those checks.

There is probably more reordering of branches to be done, but I don't know enough about the callers of that function to reorder safely, since the if-else branches return early and thus the cases might not be a real partition of the set of possible inputs.

### Result

The noop compile times of that large target are back to 56s.

### Acknowledgements

@ity and @stuhood for finding the regressing commit, and @cosmicexplorer and @illicitonion for helping find the root cause.