Fix errors for unreachable @rules. (#7660)
### Problem

While iterating in another branch, I noticed that a shadowed/unreachable `@rule` was not triggering an error. Most likely this issue was introduced when we began monomorphizing `@rules`.

### Solution

Fix the error for unreachable `@rules`. Additionally, backport a simplification to the `@rule` selection algorithm from #7600 which makes it significantly easier to describe.

### Result

Better error messages when `@rules` are shadowed or otherwise unreachable.