Allow V2 console rules to use colors (#7689)
### Problem
To improve the UX of V2 rules, there are certain times that we would like to use colors in the output, just as we do in V1.

However, we also need some way to respect the global flag `--no-colors` / `--colors`.

### Solution
Add safe color methods to the `Console` and `MockConsole` classes. These take a parameter for whether colors may be used, which in the former case get initialized by the global options.

Console rule users would then use colors like this: `console.write_stdout(console.green("Pretty colors!"))`.

#### Alternative API considered: stripping output
Instead of defining color methods on the `Console` class, we could have left the callers to use the `colors` library normally, then changed `Console.write_stderr()` et al. to call `colors.strip_colors()` on the payload.

This reduces the risk of people in the current approach accidentally directly using `colors` rather than the console methods. It also allows us to keep the definition of `Console` a bit smaller.

However, the idea of first adding color codes then intentionally stripping them away is awkward and it was decided that this was not worth it.

### Result
Colors can now be safely used in V2, which unblocks https://github.com/pantsbuild/pants/pull/7676.

The v2 test rule was also updated to print in red "Tests failed" upon any failures.

<img width="830" alt="Screen Shot 2019-05-09 at 8 53 41 AM" src="https://user-images.githubusercontent.com/14852634/57467745-f45cc900-7237-11e9-8bcb-6a7cf3c5ebc0.png">