Simplify test to test what we actually care about
We don't need to test things like compatibility flag winning over global constraints - that is not what this file is focused on. Instead, the main thing we need to test is that we can convert PythonTargetAdaptors to the proper format PEX is expecting.

This also works around issues with mocking the PythonSetup subsystem.