Improve V2 `./pants test` output to make it easier to parse results of multiple targets (#7676)
### Problem
When running `./pants test` via V2 with multiple arguments, the output is difficult to scan. Everything appears as the same target, and it is not clear what is being printed:

<img width="1256" alt="Screen Shot 2019-05-07 at 6 36 33 PM" src="https://user-images.githubusercontent.com/14852634/57343164-10a51c80-70f7-11e9-8fe6-3b7bc852b8ce.png">

### Solution

* If it's a failure, output stdout in red.
   * An even better solution would be to preserve the ANSI escape codes from Pytest. This does not seem possible because Pytest is executed in a subprocess and its output is piped, not printed to a terminal. See https://stackoverflow.com/questions/42589584/ansi-color-lost-when-using-python-subprocess.
* Add more whitespace between targets.
   * Makes it easier to scan.
* Specify which address the stdout corresponds to.
   * This is especially important as we add logging of stderr via https://github.com/pantsbuild/pants/issues/7388. That issue suggests using headers for `stdout` and `stderr`. It's important that we can disambiguate which address those correspond to.

#### Not done: ignoring stdout of success tests
Successful tests arguably don't need their stdout printed. Pytest for example silences stdout and stderr for successful tests.

However, we would only want to make this change if we gave users the options to choose the logging behavior, such as Bazel's https://docs.bazel.build/versions/master/command-line-reference.html#flag--test_output.

We don't want to introduce V2-only flags unless we absolutely have to, so we punt on this change for now.

### Result

<img width="1413" alt="Screen Shot 2019-05-08 at 12 05 06 PM" src="https://user-images.githubusercontent.com/14852634/57400964-86f25f00-7189-11e9-87d8-ef6f217d94a7.png">