Mark build-support Python files as Pants targets to lint build-support (#7633)
### Converting to Pants targets
Now that we have two Python scripts, and may have more scripts in the future, there is value in linting our script code.

By making these two scripts as targets, we can now use `./pants fmt`, `./pants lint`, and `./pants mypy` on the build-support folder for little cost.

The pre-commit check will check `./pants fmt` and `./pants lint` automatically on the `build-support` Python files now. It will not do the header check until we explicitly add the folder.

### Add `common.py`
We also add `common.py` to reduce duplication and simplify writing scripts.

See the `NB` in that file about why we only use the stdlib for that file.