V2 `./pants test.pytest` selects interpreter based off of compatibility constraints (#7679)
Currently the V2 task will always use the current interpreter to run Pytest, even if the targets have different requirements. Instead, we must support choosing the interpreter for the subprocess based on the targets' compatibility (which falls back to global compatibility if missing).

Note that we let Pex do the intepreter resolution for us to simplify the Pants code. All we have to do is pass the interpreter constraints to PEX, along with exposing the `--python-setup-interpreter-search-paths`, and then Pex will handle the resolution for us.