Refactor `check_banned_imports.py` to allow Python 3 files to import `subprocess` (#7706)
### Problem
Once we drop Python 2, we will start directly importing the stdlib module for `subprocess` and will stop using `subprocess32` and `pants.util.process_handler.subprocess`. However, our `check_banned_imports.py` would not allow this to happen. 

At the same time, we would like to keep checking that our Python 2 files are valid, so we don't want to remove the check outright.

### Solution
Extract out a `filter_files()` function that will remove any files without the given snippet. This allows us to identify and save in memory all of our Python 2 files, then to call `check_banned_import` on only those files.

We check for the presence of `from __future__` imports, which are only used in our codebase for Python 2 compatibility.

### Result
We can now import `subprocess` in Python 3-only files.

There is a 1-second slowdown in the performance of that script. After we finish dropping Python 2, we could reclaim this second by removing both the filter to get Python 2 only files and removing the `subprocess` check entirely.