Pipe LineOriented through the console (#7723)
## Problem
LineOriented is a nice utility to mix options with the console, but unfortunately it writes directly to std{out,err} after grabbing them from the Console. Ideally, we would like all output to go through the Console class, to make things like the UI easier to develop.

## Solution
Enhance print_std{out,err} in Console to accept a separator.
Refactor LineOriented a bit to only write to raw files when we are not writing to stdout.

## Result
There should be no visible change in behavior.