Fix 'current' platform check for python_dist() targets with C/C++ sources (#7687)
### Problem

Currently, in BuildLocalPythonDistributions, any `python_dist()` targets which depend on shared libs we compile and link from `ctypes_compatible_cpp_library()` targets are converted into platform-specific wheels, with the platform matching the current host (this is then injected into the build graph as a PythonRequirementLibrary). In ResolveRequirementsTaskBase, we attempt to check for whether any such `python_dist()` targets with native code exist, and if so, ensure we only resolve for the platform `['current']`. This check is currently broken in such a way that we never find any native code, and we always try to resolve for all the platforms in `--python-setup-platforms` (or whatever platforms may be set on individual `python_binary()` targets). If `--python-setup-platforms` has anything besides `['current']`, this resolve will fail, because we've only built our distributions with native code locally, i.e. for the current platform.

This wasn't noticed until #7618, which ended up essentially assigning `--python-setup-platforms` to `python_tests()` targets, causing the resolve to fail in our repo when running tests. We also then realized that we don't actually ever need to resolve for anything but the `'current'` platform *except* explicitly in `PythonBinaryCreate`, so we were able to remove all of this logic from `ResolveRequirementsTaskBase` and only trigger the check when it actually needs to fail.

### Solution

- Don't check the platforms and always resolve for `'current'` in `ResolveRequirementsTaskBase#resolve_requirements()`.
- Work around the issue catalyzed by #7618 by not assigning any platforms to `python_tests()` when platform checking unless the issue in #7616 is encountered specifically.
- Fix the broken platform check.
- Do a very small refactor of some of the `python_dist()` testing.

### Result

The platform checking in ResolveRequirementsTaskBase is removed entirely, only occurring during python binary creation instead. When that platform checking does occur, `python_tests()` will only be assigned platforms if another target has an explicit `platforms` setting, and `python_dist()` targets with native code will be successfully resolved by the `ResolveRequirements` task.