Use @union to make the v2 test runner generic (#7661)
### Problem

Fixing #4535 moved us closer to supporting a `Target` API for v2, but did not begin to use `@union` and `UnionRule` to make the v2 test `@console_rule` generic.

### Solution

Add `@union TestTarget`, and consume it in the v2 `test` `@console_rule`. The result is that a language implementer that wants to add support for testing a target type `MyTestTarget` can declare a `UnionRule(TestTarget, MyTestTarget)` in their registered rules, which would in turn require a declared `@rule` to produce a `TestResult` for `MyTestTarget`.

In order to use `PythonTestAdaptor` as a member of the `TestTarget` union, the bottom commit refactors the `SymbolTable` manipulation that we do to preserve the concrete `TargetAdaptor` classes.

### Result

Although we're not quite ready to "bless" `TargetAdaptor` by requiring it in the `SymbolTable` or coupling it to the v1 `Target` class, this change represents an incremental step in the direction of using (a likely renamed) `TargetAdaptor` incarnation as the v2 Target class.