Fix using sets with V2 `Struct` causing unhashable TypeError (#7747)
### Problem
When running our whole unit test suite with V2, we encounter a `TypeError` that sets are not hashable. This happens because one of the tests includes the value `'tags': {'platform_specific_behavior'}`, which gets stored in its `Struct`.

### Solution
Teach `Struct` how to safely hash `set`. We must first sort the `set` because `set`s have non-deterministic ordering.