Hash an entire json input in one call. (#7759)
### Problem

`hash_all` takes a list of strings, but was being passed a string. And thanks to python's venerable implicit conversion from string to list of (single character) strings, that ... "worked".

### Solution

Pass a list of strings.

### Result

Surprisingly little performance difference (less than 2% in this usecase), but a better example to follow in future.