Memoize graph walks on Context (#7758)
### Problem

In profiling of noop runs, `Context.targets(predicate=..)` was taking up a surprising amount of time due to not being memoized. While investigating, I also noticed that `MutableBuildGraph` was made unnecessary via #7243 and others.

### Solution

Remove `MutableBuildGraph`, inline one method that was unnecessarily abstract into `BuildGraph`, and then add support for a callback that can be used to invalidate caches of graph consumption. As the comments point out, we hope to eliminate mutability entirely for this usecase in the future. 

### Result

A 17% speedup for noop runs of larger targets.