Fix hermetic environment issues with V2 tasks (#7721)
### Problem

#### Specific test failures
Several of the `util` tests were failing due to issues with using a completely hermetic environment, as follows:
* `test_contextutil.py` assumes that `$USER` is defined in the env, which is not the case in V2.
* `test_dirutil.py` defaulted to using the ASCII codec for its encoding when writing via Python's `open()`, because the environment does not define `LC_ALL` or `LANG`.
* Any test that uses `test_base.py` ends up needing the cryptography wheel. This could not be built properly on macOS environments because the OpenSSL headers were not set up properly.
    * See https://www.pantsbuild.org/howto_contribute.html#getting-pants-source-code for the snippet we tell Pants devs to use.

#### Way to specify environment variables
The naive solution to these problems would simply grab the local environment variables for `LC_ALL` / `LANG` and `CPPFLAGS` / `LDFLAGS`. This works during local execution, but presents issues for remote execution where we want to have fine-grained control over the environment.

So, we must have some way to hardcode these environment variables via options, while providing sensible defaults that just work in the average case.

The greater design question is tracked by https://github.com/pantsbuild/pants/issues/7735.

### Solution
Introduce a new idiom of using a `Subsystem` to allow users to have fine-grained control over defining each env var, and then add `datatype`s and rules that compose the individual options into a meaningful set of options that are all required together to achieve the desired behavior.

For example, to fix the encoding problem, we need to surface both `LC_ALL` and `LANG` to the `ExecuteProcessRequest`, so, we want a simple way for the code to access those two variables together; the abstraction should be at the level of behavior, not the individual env vars. We also want users to be able to define each individual env var separately, so the external user API should be at the abstraction of an individual env var.

This does not close out https://github.com/pantsbuild/pants/issues/7735, as this design is not yet general enough for what we will likely need for future cases. Instead, we try to solve this problem with the minimum diff possible, while getting us closer to the final design.