Extract RuleGraph into a separate crate (#7767)
### Problem

The `RuleGraph` is relatively intertwined with the rest of the engine and has dependencies on the python code and the externs. This makes it more challenging to test, and its interface less clear. 

### Solution

Extract implementations of `Rule` and `DependencyKey` from the `RuleGraph`, then move `RuleGraph` to its own crate and parameterize it on a `R: Rule` type.

### Result

Adding tests for #7710 should be more straightforward.