Run majority of unit tests in CI with V2 Pytest runner (#7724)
### Problem
Beyond wanting to move everything towards V2 in general, this change will allow us to start remoting our unit tests for greater parallelism and less dependance on Travis.

### Solution
We must make several changes to land this:
* Use `travis_wait` to avoid a timeout. See https://docs.travis-ci.com/user/common-build-problems/#build-times-out-because-no-output-was-received.
* Introduce a blacklist to run with V1. Tracked by https://github.com/pantsbuild/pants/issues/7772.
* Fix several BUILD files that were missing the proper dependencies and tags.
* Remove support for sharding. This is not yet implemented in V2.

### Result
CI is overall less useful. The unit tests now take 50 minutes to run, rather than 20. Likewise, the logging is less useful because the use of `travis_wait` means that it is not logged until the entire process completes, and currently we log `stdout` and `stderr` for successful tests, so there is more noise.

However, these costs are deemed worth it to allow us to soon use remoting.