make assorted rsc task fixes (#7742)
### Problem

The `RscCompile` task is supposed to start running the zinc compile of a target once the rsc compiles of of all the target's dependencies are complete -- this is what allows for the massive parallelism from rsc in the first place. However, currently, the execution graph has the zinc compile of a target depending on the zinc compiles of all its dependencies, failing to use the rsc output at all.

This issue covered up a separate issue, because `javac` still can't actually use rsc output yet, and that should have caused failures, but since the zinc compiles were all scheduled after other zinc compiles, the rsc output was never being used.

These issues were fixed in #7227, but the implementation there was deemed too complex and parts were reverted without enough discussion when the "mixed compile" changes landed in master.

### Solution

- Fix execution graph dependency breakages.
- Rename `rsc-then-zinc` to `rsc-and-zinc`, since they're supposed to happen in parallel.
- Default the `--workflow` option to `rsc-and-zinc`.
- Change the display of each job in the debug output displaying the dynamic execution graph (e.g. `zinc[zinc-only]`, `zinc[zinc-java]`, and `zinc[rsc-and-zinc]`).

### Result

Some bugs are fixed!