Sort entries in output zinc jars (#7834)
### Problem

The jars produced by the `zinc` `-jar` flag were not ending up sorted due to conversion to a `Map` before writing. Additionally, they contained a duplicate entry (that `findbugs` would choke on) as an artifact of how relativization was happening.

### Solution

Preserve the sorted order that is created by the `TreeSet` that we collect entries into, and do not create an entry for the "root" directory of the inputs.

### Result

Pants is able to consume the `-jar` option successfully in #7833.