Don't implicitly materialize files from hermetic compiles
This is a breaking change, but only if you've opted into hermetic
compiles. It requires that when looking up classpaths, you pass in
`context` if you want the files referenced by the `ClasspathEntry`s to
actually be on disk. It will then implicitly materialize them when the
classpath is looked up.

This avoids needing to download files from the remote execution system
if they won't actually be needed locally (i.e. if you're not running
tests which need them, or similar). This should significantly speed up
remote executed compiles, beacuse it avoids a lot of network I/O.

A few things that changed on the way here:
* `materialize_directory` now assumes that if a file already exists
  where it's materializing with the same name, that file is already
  correct. This is valid in the compile case because we use hashes in
  paths. Generally, trying to materialize on top of existing files as
  replacement is kind of weird, so I think this is fine.
* Stop checking for overlapping paths when materializing. This is needed
  beacuse we materialize all files to the build root. Given the
  behaviour of assuming files are correct if already materialized, this
  seems like a reasonable check to remove.
* test_basic_binary now runs `./pants binary`, generating an explicit
  output at a known location, rather than running `./pants compile` and
  sniffing for files in .pants.d which are no longer implicitly
  materialized. This means we're more testing the interface rather than
  implementation, as .pants.d is viewed to be a no-guarantees black box.