Use `--no-enabled-shared` for Linux wheel-builder interpreter so that the released PEX works with statically built interpreters (#7794)
### Problem
A downstream user (Twitter) is having issues consuming the Python 3 PEX because it is complaining `ImportError: libpython3.6m.so.1.0: cannot open shared object file: No such file or directory`. This is because they build a Python interpreter from source and do so statistically, meaning that they do not use `--enable-shared` when compiling so cannot use shared libraries like `libpython3.6.m.so.1.0`.

Specifically, the issue boils down to several of our wheels not being `manylinux1` compatible, which means that they use native libraries outside of the accepted set for compatibility across Linux environments. [PEP 513](https://www.python.org/dev/peps/pep-0513/#libpythonx-y-so-1) states that it is a `manylinux1` wheel cannot rely on `libpython.X.Y.so.1`, as several of ours are doing, because statically built interpreters will fail to work.

Ideally users will be able to use our PEX regardless of whether their interpreter is built statically or not.

### Solution
No longer use `--enable-shared` for our build, which means that our wheels will not be able to depend on shared libraries, so downstream users will no longer be hit by this issue.