Run contrib tests with unit tests and integration tests for fewer CI delays (#7709)
### Problem
The contrib shards consistently are the blocker for CI runs because they appear near the end of the shards.

While discussing options to fix this, we realized that there is no compelling reason for the contrib folder to have its own dedicated shard(s). Instead, it would be valuable to use the split between unit tests vs. integration tests with the contrib code too.

### Solution
Merge contrib tests into the unit tests and integration tests shards.

### Result
The unit tests and integration tests shards will take slightly longer, but we will no longer be blocked by contrib at the end of the run.