Move V2 test runner integration test into proper location of `backend/python` folder (#7847)
`test_test_integration.py` was really an integration test for the V2 Pytest runner, not for the generic V2 test runner. For example, these tests would not make sense in the context of running Java tests. Related code should live near each other to make things more discoverable.

Also, this changes the `BUILD` for `pants_test/rules` to create an entry for each test file and to properly include each test's dependencies, which were quite out of date. Having a target per test file makes it easier to keep dependencies up to date and makes it easier to run a specific test suite.

This is prework for https://github.com/pantsbuild/pants/pull/7846.